third_party/blink/web_tests/fast/xsl/sort-locale-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
en
peach
péché
pêche
sin
fr
peach
péché
pêche
sin
fr-CA
peach
pêche
péché
sin
